+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/common/RecordStatus.java
@@ -7,5 +7,5 @@ public class RecordStatus {
 
   public AssetStatus type;
 
-  public BigInteger value;
+  public int value;
 }
