+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/info/DBInfo.java
@@ -9,11 +9,11 @@ public class DBInfo {
   public DBDriver db;
 
   @SerializedName("conn_size")
-  public BigInteger connSize;
+  public int connSize;
 
   @SerializedName("center_id")
-  public BigInteger centerId;
+  public int centerId;
 
   @SerializedName("machine_id")
-  public BigInteger machineId;
+  public int machineId;
 }
